+++ src/Requests/PurchasePaymentTermsRequest.php
@@ -1,0 +1,10 @@
+<?php
+namespace OneCommunity\Requests;
+
+class PurchasePaymentTermsRequest extends CollectionRequest
+{
+    public function getUri(): string
+    {
+        return 'purchase_payment_terms';
+    }
+}